requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
pydantic==1.10.5
fluent.runtime==0.3.1
aiogram_media_group==0.5.1